Separate traverse to enter & leave. Fixes. All tests pass now.

+++ mesa-glsl/glslang/Public/ShaderLang.h
@@ -131,7 +131,7 @@ typedef enum {
 
 typedef struct {
     ShChangeableType type;
-    long index;
+    int index;
 } ShChangeableIndex;
 
 typedef struct {
